Migrate Background to dc_bundle
I ended up moving some of the proto files around because background didn't make sense as a modifier.
It still pulls types from the modifier which is awkward, but I didn't think it made sense to create a seperate subpackage for it
