--- crates/dc_bundle/src/legacy_definition/element/image.rs
@@ -1,33 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-use serde::{Deserialize, Serialize};
-
-/// Instead of keeping decoded images in ViewStyle objects, we keep keys to the images in the
-/// ImageContext and then fetch decoded images when rendering. This means we can serialize the
-/// whole ImageContext, and always get the right image when we render.
-#[derive(Clone, PartialEq, Eq, Hash, Serialize, Deserialize, Debug)]
-pub struct ImageKey(String);
-
-impl ImageKey {
-    pub fn key(&self) -> String {
-        self.0.clone()
-    }
-
-    pub fn new(str: String) -> ImageKey {
-        ImageKey(str)
-    }
-}
