+++ crates/figma_import/src/image_context.rs
@@ -22,7 +22,7 @@ use std::{
 use crate::error::Error;
 use crate::fetch::ProxyConfig;
 use crate::figma_schema::{Paint, Transform};
-use dc_bundle::legacy_definition::element::image::ImageKey;
+use dc_bundle::legacy_definition::element::background::ImageKey;
 use image::DynamicImage;
 use serde::{Deserialize, Serialize};
 use std::hash::Hash;
