+++ crates/figma_import/src/lib.rs
@@ -46,8 +46,8 @@ pub use fetch::{fetch_doc, ConvertRequest, ConvertResponse, ProxyConfig};
 pub use image_context::ImageContextSession;
 pub use toolkit_schema::{View, ViewData}; // ugly hack
                                           // Internal convenience
+pub use dc_bundle::legacy_definition::element::background::ImageKey;
 pub use dc_bundle::legacy_definition::element::color::Color;
-pub use dc_bundle::legacy_definition::element::image::ImageKey;
 pub use dc_bundle::legacy_definition::element::node::NodeQuery;
 
 #[cfg(feature = "http_mock")]
