+++ crates/figma_import/src/reflection.rs
@@ -29,10 +29,10 @@ pub fn registry() -> serde_reflection::Result<serde_reflection::Registry> {
         .trace_type::<dc_bundle::legacy_definition::layout::positioning::AlignSelf>(&samples)
         .expect("couldn't trace AlignSelf");
     tracer
-        .trace_type::<crate::toolkit_style::ScaleMode>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::background::ScaleMode>(&samples)
         .expect("couldn't trace ScaleMode");
     tracer
-        .trace_type::<crate::toolkit_style::Background>(&samples)
+        .trace_type::<dc_bundle::legacy_definition::element::background::Background>(&samples)
         .expect("couldn't trace Background");
     tracer
         .trace_type::<dc_bundle::legacy_definition::modifier::blend::BlendMode>(&samples)
