+++ crates/figma_import/src/vector_schema.rs
@@ -14,12 +14,13 @@
 
 use std::convert::TryFrom;
 
+use dc_bundle::legacy_definition::element::background::Background;
 use dc_bundle::legacy_definition::modifier::blend::BlendMode;
 use dc_bundle::legacy_definition::modifier::shadow::BoxShadow;
 use dc_bundle::legacy_definition::modifier::transform::AffineTransform;
 use serde::{Deserialize, Serialize};
 
-use crate::toolkit_style::{Background, Stroke};
+use crate::toolkit_style::Stroke;
 
 // This is a simple vector representation that should reproduce Figma or
 // another design tool's rendering. It simplifies the renderer's task by
