+++ proto/definition/element/path.proto
@@ -16,7 +16,7 @@ syntax = "proto3";
 
 package designcompose.definition.element;
 
-import "definition/modifier/background.proto";
+import "definition/element/background.proto";
 
 option java_package = "com.android.designcompose.definition.element";
 
@@ -89,5 +89,5 @@ message StrokeWeight {
 message Stroke {
   StrokeAlign stroke_align = 1;
   StrokeWeight stroke_weight = 2;
-  repeated modifier.Background strokes = 3;
+  repeated Background strokes = 3;
 }
